Simplify NullAuthenticator, add test
